backend/requirements.txt
# FastAPI 核心
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# 数据库
sqlalchemy==2.0.35
psycopg[binary]==3.2.10
alembic==1.13.3

# 数据处理
pandas==2.2.3
numpy==2.1.2
openpyxl==3.1.5

# 定时任务
apscheduler==3.10.4

# AI & HTTP
httpx==0.27.2
openai==1.51.2

# 报告导出
weasyprint==62.3
python-docx==1.1.2
jinja2==3.1.4

# 飞书SDK
lark-oapi==1.2.19

# 工具
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2